Salvaged Code
The code in "Snap! Raise_Fundraising for Teams, Groups & Clubs (1).html" was salvaged by downloading a local copy of Snap Raise's homepage and adding modifications to demo the new map feature as necessary. The code responsible for the new map feature is primarily found in "mapbox_local.js" (optimized for local feature demo in Google Chrome) and "_landing_hero.css" (originally written as a .scss file).